func:policies create
policies are created and implemented

+++ app/Http/Controllers/JobController.php
@@ -7,9 +7,12 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -97,6 +100,9 @@ class JobController extends Controller
     //@route GET /jobs/{$id}/edit
     public function edit(Job $job): View
     {
+        //Check if user authorized
+        $this->authorize('update', $job);
+
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -107,6 +113,9 @@ class JobController extends Controller
     //@route PUT /jobs/{$id}
     public function update(Request $request, Job $job): string
     {
+        //Check if user authorized
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -155,6 +164,9 @@ class JobController extends Controller
     //@route DELETE /jobs/{$id}
     public function destroy(Job $job): RedirectResponse
     {
+        //Check if user authorized
+        $this->authorize('delete', $job);
+
         //If logo then delete it
         if ($job->company_logo) {
             Storage::delete('public/logos/' . $job->company_logo);
